Move parallel phase tests into a separate file
